+++ addon/components/paper-dialog-content.js
@@ -2,9 +2,7 @@
  * @module ember-paper
  */
 import Ember from 'ember';
-import PaperDialogInner from './paper-dialog-inner';
-
-const { Component, computed, run } = Ember;
+const { Component } = Ember;
 
 /**
  * @class PaperDialogComponent
@@ -12,30 +10,5 @@ const { Component, computed, run } = Ember;
  */
 export default Component.extend({
   tagName: 'md-dialog-content',
-  classNames: ['md-dialog-content'],
-
-  dialogInnerComponent: computed(function() {
-    return this.nearestOfType(PaperDialogInner);
-  }),
-
-  imagesLoaded() {
-    let content = this.get('element');
-    this.get('dialogInnerComponent').set('contentOverflow', content.scrollHeight > content.clientHeight);
-  },
-
-  didInsertElement() {
-    // content overflow might change depending on load of images inside dialog.
-    let images = this.$().find('img');
-
-    let imageLoadHandler = run.bind(this, this.imagesLoaded);
-    this.set('imageLoadHandler', imageLoadHandler);
-
-    images.on('load', imageLoadHandler);
-  },
-
-  willDestroyElement() {
-    this._super(...arguments);
-    let images = this.$().find('img');
-    images.off('load', this.get('imageLoadHandler'));
-  }
+  classNames: ['md-dialog-content']
 });