+++ addon/components/paper-dialog-inner.js
@@ -4,7 +4,7 @@
 import Ember from 'ember';
 import Translate3dMixin from '../mixins/translate3d-mixin';
 
-const { Component } = Ember;
+const { Component, run } = Ember;
 
 /**
  * @class PaperDialogInner
@@ -30,6 +30,23 @@ export default Component.extend(Translate3dMixin, {
     if ($origin) {
       $origin.focus();
     }
-  }
+  },
 
+  imagesLoaded() {
+    let content = this.element.querySelector('md-dialog-content');
+    this.set('contentOverflow', content.scrollHeight > content.clientHeight);
+  },
+
+  didInsertElement() {
+    this._super(...arguments);
+    // content overflow might change depending on load of images inside dialog.
+    let images = this.$().find('img');
+    images.on(`load.${this.elementId}`, run.bind(this, this.imagesLoaded));
+  },
+
+  willDestroyElement() {
+    this._super(...arguments);
+    let images = this.$().find('img');
+    images.off(`load.${this.elementId}`);
+  }
 });